- Update to how 'iocage fetch' works.
If the user supplies a release=<foo>, we don't print the screen
asking which release to use, since they already indicated this.

If the initial "fetch" fails of the dist-files, we will bail out.

When exporting user-supplied values, we will also export iocset_$val
so we can parse if we are using a default value, or one the user
explicitly set via args.