Limit scope of IFS change in ioc-network
Fixes a bug affecting the shutdown of jails with shared IP networking. Setting `IFS=','` sticks for the rest of the script after the first call to `__stop_legacy_networking`, and causes many subsequent functions to misbehave. Limiting `IFS` to local scope ensures the rest of the script is not affected.