Create helpers for getting a property from a dataset.
Internally we usually know the dataset for a jail and an be sure the property
we are getting is supported, so the lookups and checks in __get_jail_prop are
excessive. The helpers __get_dataset_zfs_prop and __get_dataset_zfs_prop
can be used to replace most occurrences of __get_jail_prop, and doing so should
greatly improve the performance of iocage.