Shorten list of jails to start.
Second try.
If a iocage server hold backups of other servers the server will try to
start them all. Even if we know they won’t start.
This change create a list of jails that match the hosted of the host
and the jail. If they match the jail is started, else it discards the
jail.