Fix destroy seeing the wrong jail when using a tag that's been overwritten.
Basejails were getting created with properties on both the `iocage/jails/UUID` and `iocage/jails/UUID/root` datasets. This is not the case now, and all `/root` datasets are ignored. In addition we are setting any inheritable property during cloning to the value of `-`. This helps alleviate the problem as well.