Allow templates to be force started, and custom jail datasets.
The user can now do a `iocage start -f` for a template to allow for live editing of it. The template needs to be stopped for these changes to propogate to children.
Secondly, you can now use custom jailed datasets (briefly discussed in https://github.com/iocage/iocage/issues/271). Before we restricted you to only using `UUID/data`.
The user is also now able to change the jailed dataset mount inside the jail without restarting. This change also brings along with it another fun detail, and that is custom user mounts.

The user simply needs to create a directory in the jail as if it was a read-write mount. This requires no extra effort for the user. You can also create a directory in `/iocage/jails/UUID/_` and it will be there in the jail.