ioc-image: fix iocage import for images
iocage import was broken for images. If the UUID given
did not match a package, we tried to compute a sha256
on an empty pathname in $_package and it blocked forever
on stdin.

The code in question looked like it belonged inside
the if block for $_pcount > 0 but might have been
mistakenly moved to the start of the __import() func.

This patch puts the checks where they belong.

Signed-off-by: William Katsak <wkatsak@gmail.com>